requirements.txt
aiohttp==3.7.4.post0
aiohttp-apispec==2.2.1
aiohttp-session==2.9.0
aioresponses==0.7.2
alembic==1.6.5
apispec==3.3.2
appdirs==1.4.4
async-timeout==3.0.1
asyncpg==0.23.0
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
black==21.7b0
cffi==1.15.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coverage<=6.3.2
cryptography==3.3.2
freezegun==1.1.0
gino==1.0.1
idna==3.2
iniconfig==1.1.1
Jinja2==2.11.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-dataclass==8.5.3
multidict==5.1.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.11.0
pycparser==2.21
pyparsing==3.0.7
pytest==6.2.4
pytest-aiohttp==0.3.0
pytest-cov==2.12.1
pytest-mock==3.6.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
PyYAML==5.4.1
regex==2021.7.6
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
toml==0.10.2
tomli==1.1.0
typing-extensions==3.10.0.0
typing-inspect==0.7.1
webargs==5.5.3
yarl==1.6.3
